@@ -3,10 +3,10 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>me.mayo</groupId>
     <artifactId>TelnetKek</artifactId>
-    <version>1.0</version>
+    <version>1.1</version>
     <packaging>jar</packaging>
     <build>
-	<finalName>do-not-use-this</finalName>
+        <finalName>do-not-use-this</finalName>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -63,8 +63,8 @@
     </dependencies>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.7</maven.compiler.source>
-        <maven.compiler.target>1.7</maven.compiler.target>
+        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>1.8</maven.compiler.target>
     </properties>
     <name>TelnetKek</name>
 </project>
